% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsod.R
\name{fsod}
\alias{fsod}
\title{irst-Second & Odd-Even Split Half}
\usage{
fsod(df.split, method, sub, vars)
}
\arguments{
\item{df.split}{A list of split datasets}

\item{method}{The method used for splitting}

\item{sub}{Subject identifier}

\item{vars}{Variables to be considered}
}
\value{
Returns the result of the split
}
\description{
irst-Second & Odd-Even Split Half
}
