% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{Monte-Carlo Split Half}
\usage{
mc(df.split, iteration, nc, sub, vars)
}
\arguments{
\item{df.split}{A list of split datasets}

\item{iteration}{Number of iterations}

\item{nc}{Number of CPU threads to use}

\item{sub}{Column indicating the subject}

\item{vars}{Variables to be considered}
}
\value{
Returns the result of the Monte Carlo split
}
\description{
Monte-Carlo Split Half
}
