% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_tool.R
\name{sh_tool}
\alias{sh_tool}
\title{Function for Splitting Cognitive Experiment Datasets}
\usage{
sh_tool(df.split, iteration, nc, sub, vars, method)
}
\arguments{
\item{df.split}{Dataset to be split}

\item{iteration}{Number of iterations. Required only for Monte Carlo and Permuted split half}

\item{nc}{nunmber of cores Required only for Monte Carlo and Permuted split half}

\item{sub}{Subject ID}

\item{vars}{Stratification variables, i.e., all variables in your study}

\item{method}{Splitting method, choose one from the following four: 'mc', 'fs', 'od', 'permuted}
}
\value{
Returns the result of the split
}
\description{
Function for Splitting Cognitive Experiment Datasets
}
